+++ src/main/java/com/followMe/delivery_server/delivery/application/ShipmentServiceImpl.java
@@ -6,7 +6,7 @@ import static com.followMe.delivery_server.delivery.application.DeliveryPermissi
 import com.followMe.delivery_server.delivery.application.dto.DeliveryResponseDto;
 import com.followMe.delivery_server.delivery.domain.Shipment;
 import com.followMe.delivery_server.delivery.domain.enums.ShipmentStatus;
-import com.followMe.delivery_server.delivery.domain.exception.DeliveryException;
+import com.followMe.delivery_server.delivery.domain.exception.ShipmentNotFoundException;
 import com.followMe.delivery_server.delivery.domain.repository.DeliveryRepository;
 import com.followMe.delivery_server.delivery.infra.query.DeliveryQueryRepository;
 import java.util.List;
@@ -48,9 +48,7 @@ public class ShipmentServiceImpl implements ShipmentService {
   @Transactional
   public void updateStatus(UserContext user, UUID shipmentId, ShipmentStatus status) {
     Shipment shipment =
-        deliveryRepository
-            .findShipmentById(shipmentId)
-            .orElseThrow(DeliveryException.ShipmentNotFoundException::new);
+        deliveryRepository.findShipmentById(shipmentId).orElseThrow(ShipmentNotFoundException::new);
     checkShipmentStatusUpdateAccess(user, shipment);
     shipment.updateShipmentStatus(status);
   }
