+++ src/main/java/com/followMe/delivery_server/delivery/application/UserRole.java
@@ -4,5 +4,5 @@ public enum UserRole {
   VENDOR,
   DELIVERY_MANAGER,
   HUB_MANAGER,
-  MASTER;
+  MASTER
 }
