+++ src/main/java/com/followMe/delivery_server/delivery/domain/exception/DeliveryNotFoundException.java
@@ -1,0 +1,9 @@
+package com.followMe.delivery_server.delivery.domain.exception;
+
+import com.followMe.common.exception.BusinessException;
+
+public class DeliveryNotFoundException extends BusinessException {
+  public DeliveryNotFoundException() {
+    super(DeliveryErrorCode.DELIVERY_NOT_FOUND);
+  }
+}
