+++ src/main/java/com/followMe/delivery_server/delivery/domain/exception/HubClientUnavailableException.java
@@ -1,0 +1,9 @@
+package com.followMe.delivery_server.delivery.domain.exception;
+
+import com.followMe.common.exception.BusinessException;
+
+public class HubClientUnavailableException extends BusinessException {
+  public HubClientUnavailableException() {
+    super(DeliveryErrorCode.HUB_CLIENT_UNAVAILABLE);
+  }
+}
