+++ src/main/java/com/followMe/delivery_server/delivery/domain/exception/HubRouteNotFoundException.java
@@ -1,0 +1,9 @@
+package com.followMe.delivery_server.delivery.domain.exception;
+
+import com.followMe.common.exception.BusinessException;
+
+public class HubRouteNotFoundException extends BusinessException {
+  public HubRouteNotFoundException() {
+    super(DeliveryErrorCode.HUB_ROUTE_NOT_FOUND);
+  }
+}
