+++ src/main/java/com/followMe/delivery_server/delivery/domain/exception/InvalidNodeInformationException.java
@@ -1,0 +1,9 @@
+package com.followMe.delivery_server.delivery.domain.exception;
+
+import com.followMe.common.exception.BusinessException;
+
+public class InvalidNodeInformationException extends BusinessException {
+  public InvalidNodeInformationException() {
+    super(DeliveryErrorCode.INVALID_NODE_INFORMATION);
+  }
+}
