+++ src/main/java/com/followMe/delivery_server/delivery/domain/exception/NodeTypeMismatchException.java
@@ -1,0 +1,9 @@
+package com.followMe.delivery_server.delivery.domain.exception;
+
+import com.followMe.common.exception.BusinessException;
+
+public class NodeTypeMismatchException extends BusinessException {
+  public NodeTypeMismatchException() {
+    super(DeliveryErrorCode.NODE_TYPE_MISMATCH);
+  }
+}
