+++ src/main/java/com/followMe/delivery_server/delivery/domain/exception/ShipmentNotFoundException.java
@@ -1,0 +1,9 @@
+package com.followMe.delivery_server.delivery.domain.exception;
+
+import com.followMe.common.exception.BusinessException;
+
+public class ShipmentNotFoundException extends BusinessException {
+  public ShipmentNotFoundException() {
+    super(DeliveryErrorCode.SHIPMENT_NOT_FOUND);
+  }
+}
