+++ src/main/java/com/followMe/delivery_server/delivery/infra/query/DeliveryQueryRepository.java
@@ -8,8 +8,8 @@ import com.followMe.common.pagination.PageResponse;
 import com.followMe.delivery_server.delivery.application.dto.DeliveryListItemDto;
 import com.followMe.delivery_server.delivery.application.dto.DeliveryResponseDto;
 import com.followMe.delivery_server.delivery.application.dto.DeliverySearchCondition;
-import com.followMe.delivery_server.delivery.domain.exception.DeliveryException;
-import com.followMe.delivery_server.delivery.domain.exception.DeliveryException.DeliveryNotFoundException;
+import com.followMe.delivery_server.delivery.domain.exception.DeliveryNotFoundException;
+import com.followMe.delivery_server.delivery.domain.exception.ShipmentNotFoundException;
 import java.util.List;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
@@ -96,13 +96,13 @@ public class DeliveryQueryRepository {
             .where(P_SHIPMENT.DELIVERY_ID.eq(deliveryId))
             .orderBy(P_SHIPMENT.SEQUENCE.asc())
             .fetch();
-    if (records.isEmpty()) throw new DeliveryException.ShipmentNotFoundException();
+    if (records.isEmpty()) throw new ShipmentNotFoundException();
     return recordMapper.toShipmentResponseList(records);
   }
 
   public DeliveryResponseDto.ShipmentResponse findShipmentById(UUID shipmentId) {
     Record record = dsl.select().from(P_SHIPMENT).where(P_SHIPMENT.ID.eq(shipmentId)).fetchOne();
-    if (record == null) throw new DeliveryException.ShipmentNotFoundException();
+    if (record == null) throw new ShipmentNotFoundException();
     return recordMapper.toShipmentResponse(record);
   }
 }
